simplify implementation and make two versions
